refactor: service-level aggregation
