+++ simaple/request/adapter/ability_loader/_schema.py
@@ -1,0 +1,15 @@
+from typing import TypedDict
+
+
+class _CharacterAbilityLineResponse(TypedDict):
+    ability_no: str
+    ability_grade: str
+    ability_value: str
+
+
+class CharacterAbilityResponse(TypedDict):
+    date: str
+    ability_grade: str
+    ability_info: list[_CharacterAbilityLineResponse]
+    remain_fame: float
+    preset_no: int
