+++ simaple/request/adapter/propensity_loader/_schema.py
@@ -1,0 +1,11 @@
+from typing import TypedDict
+
+
+class CharacterPropensityResponse(TypedDict):
+    date: str
+    charisma_level: int
+    sensibility_level: int
+    insight_level: int
+    willingness_level: int
+    handicraft_level: int
+    charm_level: int
