+++ simaple/request/adapter/translator/base.py
